Move store to where it's needed
