+++ src/stores/useStore.js
@@ -31,7 +31,6 @@ const useStore = create( set => ( {
   evidenceToAdd: [],
   galleryUris: [],
   groupedPhotos: [],
-  loading: false,
   observations: [],
   originalCameraUrisMap: {},
   photoEvidenceUris: [],
@@ -55,7 +54,6 @@ const useStore = create( set => ( {
     evidenceToAdd: [],
     galleryUris: [],
     groupedPhotos: [],
-    loading: false,
     observations: [],
     originalCameraUrisMap: {},
     photoEvidenceUris: [],
